+++ addon/mixins/actions-for-primitives/fd-actions-for-sd-primitives.js
@@ -4,13 +4,13 @@ import fdSequenceActor from 'ember-flexberry-designer/objects/uml-primitives/fd-
 import fdSequenceDiagramObject from 'ember-flexberry-designer/objects/uml-primitives/fd-uml-sequence-object';
 import fdSequenceDiagramActiveObject from 'ember-flexberry-designer/objects/uml-primitives/fd-uml-sequence-active-object';
 import fdInScope from 'ember-flexberry-designer/objects/uml-primitives/fd-uml-sequence-in-scope';
-import { Terminator } from '../../objects/uml-primitives/fd-uml-terminator';
-import { ProcedureCall } from '../../objects/uml-primitives/fd-uml-procedure-call';
-import { FlatMessage } from '../../objects/uml-primitives/fd-uml-flat-message';
-import { AsyncMessage } from '../../objects/uml-primitives/fd-uml-async-message';
-import { ReturnMessage } from '../../objects/uml-primitives/fd-uml-return-message';
+import Terminator from '../../objects/uml-primitives/fd-uml-terminator';
+import ProcedureCall from '../../objects/uml-primitives/fd-uml-procedure-call';
+import FlatMessage from '../../objects/uml-primitives/fd-uml-flat-message';
+import AsyncMessage from '../../objects/uml-primitives/fd-uml-async-message';
+import ReturnMessage from '../../objects/uml-primitives/fd-uml-return-message';
 import TimeConstraint from '../../objects/uml-primitives/fd-uml-time-constraint';
-import { getJsonForElement } from '../../utils/get-json-for-diagram';
+import { getJsonForElement, getJsonForLink } from '../../utils/get-json-for-diagram';
 
 /**
   Actions for creating joint js elements on cad diagrams.
@@ -87,13 +87,21 @@ export default Mixin.create({
       @param {jQuery.Event} e event.
      */
     addTerminator(e) {
-      this.createObjectData((function(x, y) {
-        let newTerminatorObject = new Terminator({
-          position: { x: x, y: y }
-        });
+      this.createObjectData((function(x, y, parentPrimitive) {
+        let jsonObject = getJsonForElement(
+          'STORMCASE.UML.sd.Terminator, UMLSD',
+          { x, y },
+          { width: 20, height: 20 },
+          { Name: '' },
+          { ConnectedPrimitive : { $ref: parentPrimitive } }
+        );
 
-        return newTerminatorObject;
-      }).bind(this), e);
+        let terminatorObject = Terminator.create({ primitive: jsonObject });
+
+        this._addToPrimitives(terminatorObject);
+
+        return terminatorObject.JointJS();
+      }).bind(this), e, A(['flexberry.uml.sequencediagramActiveObject', 'flexberry.uml.sequencediagramObject']));
     },
 
     /**
@@ -104,17 +112,20 @@ export default Mixin.create({
      */
     addProcedureCall(e) {
       this.createLinkData((function(linkProperties) {
-        let newProcedureCallObject = new ProcedureCall({
-          source: {
-            id: linkProperties.source
-          },
-          target: {
-            id: linkProperties.target
-          },
-          vertices: linkProperties.points || A()
-        });
-
-        return newProcedureCallObject;
+        let jsonObject = getJsonForLink(
+          'STORMCASE.UML.sd.ProcedureCall, UMLSD',
+          linkProperties.source,
+          null,
+          linkProperties.target,
+          null,
+          A(),
+          { Name: '', LeftText: '', RightText: '' },
+          { NamePos: 0.0, InitialMultiplicity: 1.0 }
+        );
+        let newProcedureCallObject = ProcedureCall.create({ primitive: jsonObject });
+        newProcedureCallObject.set('vertices', linkProperties.points || A());
+        this._addToPrimitives(newProcedureCallObject);
+        return newProcedureCallObject.JointJS();
       }).bind(this), e, A(['flexberry.uml.sequencediagramActiveObject', 'flexberry.uml.sequencediagramObject', 'flexberry.uml.SequenceDiagramActor']));
     },
 
@@ -126,15 +137,23 @@ export default Mixin.create({
      */
     addFlatMessage(e) {
       this.createLinkData((function(linkProperties) {
-        let newFlatMessageObject = new FlatMessage({
-          source: {
-            id: linkProperties.source
-          },
-          target: {
-            id: linkProperties.target
-          },
-          vertices: linkProperties.points || A()
-        });
+        let jsonObject = getJsonForLink(
+          'STORMCASE.UML.sd.FlatMessage, UMLSD',
+          linkProperties.source,
+          null,
+          linkProperties.target,
+          null,
+          A(),
+          { Name: '', LeftText: '', RightText: '' },
+          { NamePos: 0.0, InitialMultiplicity: 1.0 }
+        );
+
+        let flatMessageObject = FlatMessage.create({ primitive: jsonObject });
+
+        flatMessageObject.set('vertices', linkProperties.points || A());
+        this._addToPrimitives(flatMessageObject);
+
+        let newFlatMessageObject = flatMessageObject.JointJS();
 
         return newFlatMessageObject;
       }).bind(this), e, A(['flexberry.uml.sequencediagramActiveObject', 'flexberry.uml.sequencediagramObject', 'flexberry.uml.SequenceDiagramActor']));
@@ -148,15 +167,23 @@ export default Mixin.create({
      */
     addAsyncMessage(e) {
       this.createLinkData((function(linkProperties) {
-        let newAsyncMessageObject = new AsyncMessage({
-          source: {
-            id: linkProperties.source
-          },
-          target: {
-            id: linkProperties.target
-          },
-          vertices: linkProperties.points || A()
-        });
+        let jsonObject = getJsonForLink(
+          'STORMCASE.UML.sd.AsyncMessage, UMLSD',
+          linkProperties.source,
+          null,
+          linkProperties.target,
+          null,
+          A(),
+          { Name: '', LeftText: '', RightText: '' },
+          { NamePos: 0.0, InitialMultiplicity: 1.0 }
+        );
+
+        let asyncMessageObject = AsyncMessage.create({ primitive: jsonObject });
+
+        asyncMessageObject.set('vertices', linkProperties.points || A());
+        this._addToPrimitives(asyncMessageObject);
+
+        let newAsyncMessageObject = asyncMessageObject.JointJS();
 
         return newAsyncMessageObject;
       }).bind(this), e, A(['flexberry.uml.sequencediagramActiveObject', 'flexberry.uml.sequencediagramObject', 'flexberry.uml.SequenceDiagramActor']));
@@ -170,15 +197,23 @@ export default Mixin.create({
      */
     addReturnMessage(e) {
       this.createLinkData((function(linkProperties) {
-        let newReturnMessageObject = new ReturnMessage({
-          source: {
-            id: linkProperties.source
-          },
-          target: {
-            id: linkProperties.target
-          },
-          vertices: linkProperties.points || A()
-        });
+        let jsonObject = getJsonForLink(
+          'STORMCASE.UML.sd.ReturnMessage, UMLSD',
+          linkProperties.source,
+          null,
+          linkProperties.target,
+          null,
+          A(),
+          { Name: '', LeftText: '', RightText: '' },
+          { NamePos: 0.0, InitialMultiplicity: 1.0 }
+        );
+
+        let returnMessageObject = ReturnMessage.create({ primitive: jsonObject });
+
+        returnMessageObject.set('vertices', linkProperties.points || A());
+        this._addToPrimitives(returnMessageObject);
+
+        let newReturnMessageObject = returnMessageObject.JointJS();
 
         return newReturnMessageObject;
       }).bind(this), e, A(['flexberry.uml.sequencediagramActiveObject', 'flexberry.uml.sequencediagramObject', 'flexberry.uml.SequenceDiagramActor']));