+++ addon/objects/uml-primitives/fd-uml-async-message.js
@@ -5,8 +5,8 @@ import { computed } from '@ember/object'
 import joint from 'npm:jointjs';
 
 import FdUmlLink from './fd-uml-link';
-import { FlatMessage } from './fd-uml-flat-message';
-import { EmptyView } from './links-view/fd-empty-view';
+import { Link } from './fd-uml-link';
+import { RoleView } from './links-view/fd-role-view';
 
 /**
   An object that describes a Async Message on the UML diagram.
@@ -15,7 +15,6 @@ import { EmptyView } from './links-view/fd-empty-view';
   @extends FdUmlLink
 */
 export default FdUmlLink.extend({
-
   /**
     End role text.
 
@@ -39,9 +38,19 @@ export default FdUmlLink.extend({
     @method JointJS
   */
   JointJS() {
-    let properties = this.getProperties('id', 'source', 'target', 'labels');
+    let properties = this.getProperties('id', 'source', 'target', 'vertices', 'labels');
     properties.objectModel = this;
+    properties.centredAnchor = true;
     return new AsyncMessage(properties);
+  },
+}, {
+  getLabelDistance: function (labelName) {
+    switch (labelName) {
+      case 'endRole':
+        return -30;
+      case 'description':
+        return 0.5;
+    }
   },
 });
 
@@ -50,52 +59,17 @@ export default FdUmlLink.extend({
 
   @for FdUmlAsyncMsg
   @class AsyncMessage
-  @extends flexberry.uml.FlatMessage
+  @extends flexberry.uml.Link
   @namespace flexberry.uml
   @constructor
 */
-export let AsyncMessage = FlatMessage.define('flexberry.uml.sequencediagramAsyncMessage', {
+export let AsyncMessage = Link.define('flexberry.uml.sequencediagramAsyncMessage', {
   attrs: {
-    '.marker-source': { d: 'M 0 10 L 13 17 L 0 10 L 13 3 z', fill: 'black' },
-    '.connection': { stroke: 'black', 'stroke-width': 1 }
+    '.marker-target': { d: 'M 0 10 L 13 17 M 13 3', fill: 'black' },
+    '.connection': { stroke: 'black', 'stroke-width': 1 },
+    text: { visibility: 'hidden' },
+    rect: { visibility: 'hidden' }
   },
-  labels: [{
-    position: { distance: 0.05, offset: -12 }, attrs: { text: { text: '' } } //endMultiplicity
-  }, {
-    position: { distance: 0.95, offset: -12 }, attrs: { text: { text: '' } } //startMultiplicity
-  }, {
-    textAnchor: 'middle', attrs: { text: { text: '' } } //description
-  }, {
-    position: { distance: 0.05, offset: 12 }, attrs: { text: { text: '' } } //endRoleTxt
-  }, {
-    position: { distance: 0.95, offset: 12 }, attrs: { text: { text: '' } } //startRoleTxt
-  }]
-}, {
-  setLabelText: function (label, text) {
-    switch (label) {
-      case 'startMultiplicity':
-        this.label(0, { attrs: { text: { text: text } } });
-        break;
-      case 'startRole':
-        this.label(1, { attrs: { text: { text: text } } });
-        break;
-      case 'description':
-        this.label(2, { attrs: { text: { text: text } } });
-        break;
-      case 'endRole':
-        this.label(3, { attrs: { text: { text: text } } });
-        break;
-      case 'endMultiplicity':
-        this.label(4, { attrs: { text: { text: text } } });
-        break;
-      default:
-        // eslint-disable-next-line no-console
-        console.log('ERROR - choose correct label name');
-        break;
-    }
-
-    return;
-  }
 });
 
-joint.shapes.flexberry.uml.AsyncMessageView = EmptyView;
+joint.shapes.flexberry.uml.sequencediagramAsyncMessageView = RoleView;