+++ CHANGELOG.md
@@ -4,14 +4,19 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
- * Added text wrapping to the next line when entering text into a transition element in a activity diagram
- 
+* Added text wrapping to the next line when entering text into a transition element in a activity diagram
+* Added a procedure call link to a sequence diagram
+
 ### Fixed
  * Fix for error in activity diagram when editing links
  * Fix for saving error in collaboration diagram
  * Fixed default size for Object and Active Object on sequence diagram 
  * Fixed In Scope on sequence diagram
  * Fixed UML Primitive behavior when resizing
+ * Added Terminator element to Sequence Diagram.
+ * Added Async Message link element to Sequence Diagram
+ * Added Flat Message link element to Sequence Diagram
+ * Added Return Message link element to Sequence Diagram
 
 ## [1.1.2-beta.3] - 2025-03-10
 ### Fixed
@@ -19,6 +24,7 @@ All notable changes to this project will be documented in this file.
  * Changed loading detail with head flag to true at view editing.
  * Fixed selecting the same stereotype twice
  * Fixed drawing of links in Firefox
+ * Added text wrapping to the next line when entering text into a transition element in a state diagram and a connection element in a deployment diagram
  * Fixed pointer down on usecase actor.
  * Removed underline from labels in deployment diagram
  * Fixed Actor and Object on sequence diagram